test: migrate app test
