+++ test/integration/create-next-app/utils.ts
@@ -1,7 +1,6 @@
 import execa from 'execa'
-import resolveFrom from 'resolve-from'
-import { realpath } from 'fs-extra'
 import { fetchViaHTTP, findPort, killApp, launchApp } from 'next-test-utils'
+import { join } from 'path'
 
 export const CNA_PATH = require.resolve('create-next-app/dist/index.js')
 export const EXAMPLE_PATH =
@@ -39,33 +38,31 @@ export const command = (cmd: string, args: string[]) =>
     env: { ...process.env },
   })
 
-export const startsWithoutError = async (
-  cwd: string,
-  modes = ['default', 'turbo'],
-  usingAppDirectory: boolean = false
-) => {
-  for (const mode of modes) {
-    cwd = await realpath(cwd)
-    const appPort = await findPort()
-    const app = await launchApp(cwd, appPort, {
-      turbo: mode === 'turbo',
-      cwd: cwd,
-      nextBin: resolveFrom(cwd, 'next/dist/bin/next'),
-    })
+export async function tryNextDev({
+  cwd,
+  projectName,
+  isApp = true,
+}: {
+  cwd: string
+  projectName: string
+  isApp?: boolean
+}) {
+  const dir = join(cwd, projectName)
+  const port = await findPort()
+  const app = await launchApp(dir, port)
 
-    try {
-      const res = await fetchViaHTTP(appPort, '/')
-      expect(await res.text()).toContain('Get started by editing')
-      expect(res.status).toBe(200)
+  try {
+    const res = await fetchViaHTTP(port, '/')
+    expect(await res.text()).toContain('Get started by editing')
+    expect(res.status).toBe(200)
 
-      if (!usingAppDirectory) {
-        const apiRes = await fetchViaHTTP(appPort, '/api/hello')
-        expect(await apiRes.json()).toEqual({ name: 'John Doe' })
-        expect(apiRes.status).toBe(200)
-      }
-    } finally {
-      await killApp(app)
+    if (!isApp) {
+      const apiRes = await fetchViaHTTP(port, '/api/hello')
+      expect(await apiRes.json()).toEqual({ name: 'John Doe' })
+      expect(apiRes.status).toBe(200)
     }
+  } finally {
+    await killApp(app)
   }
 }
 
@@ -73,6 +70,8 @@ export {
   createNextApp,
   projectFilesShouldExist,
   shouldBeTemplateProject,
+  shouldBeJavascriptProject,
+  shouldBeTypescriptProject,
   spawnExitPromise,
 } from './lib/utils'
 export { useTempDir } from '../../lib/use-temp-dir'
